Move more things only used in one place
