+++ lib/knex-builder/internal/batch-insert.js
@@ -1,6 +1,6 @@
 const chunk = require('lodash/chunk');
 const flatten = require('lodash/flatten');
-const delay = require('../../util/delay');
+const delay = require('./delay');
 const { isNumber } = require('../../util/is');
 
 function batchInsert(client, tableName, batch, chunkSize = 1000) {
