+++ lib/schema/migrate/MigrationGenerator.js
@@ -1,7 +1,7 @@
 const path = require('path');
 const { writeJsFileUsingTemplate } = require('../util/template');
 const { getMergedConfig } = require('./configuration-merger');
-const { ensureDirectoryExists } = require('../../util/fs');
+const { ensureDirectoryExists } = require('../util/fs');
 const { yyyymmddhhmmss } = require('../util/timestamp');
 
 class MigrationGenerator {
