+++ lib/schema/migrate/sources/fs-migrations.js
@@ -1,7 +1,7 @@
 const path = require('path');
 const sortBy = require('lodash/sortBy');
 
-const { readdir } = require('../../../util/fs');
+const { readdir } = require('../../util/fs');
 
 const DEFAULT_LOAD_EXTENSIONS = Object.freeze([
   '.co',
