+++ lib/schema/seed/Seeder.js
@@ -5,10 +5,7 @@ const path = require('path');
 const flatten = require('lodash/flatten');
 const extend = require('lodash/extend');
 const includes = require('lodash/includes');
-const {
-  ensureDirectoryExists,
-  getFilepathsInFolder,
-} = require('../../util/fs');
+const { ensureDirectoryExists, getFilepathsInFolder } = require('../util/fs');
 const { writeJsFileUsingTemplate } = require('../util/template');
 const { yyyymmddhhmmss } = require('../util/timestamp');
