+++ lib/schema/util/template.js
@@ -1,6 +1,6 @@
 const template = require('lodash/template');
 
-const { readFile, writeFile } = require('../../util/fs');
+const { readFile, writeFile } = require('./fs');
 
 /**
  * Light wrapper over lodash templates making it safer to be used with javascript source code.
