+++ scripts/stress-test/knex-stress-test.js
@@ -3,7 +3,7 @@ const Knex = require('../../lib');
 const toxiproxy = require('toxiproxy-node-client');
 const toxicli = new toxiproxy.Toxiproxy('http://localhost:8474');
 const rp = require('request-promise-native');
-const delay = require('../../lib/util/delay');
+const delay = require('../../lib/knex-builder/internal/delay');
 
 // init instances
 const pg = Knex({
