+++ test/integration/builder/transaction.js
@@ -6,7 +6,7 @@ const Knex = require('../../../knex');
 const _ = require('lodash');
 const sinon = require('sinon');
 const { KnexTimeoutError } = require('../../../lib/util/timeout');
-const delay = require('../../../lib/util/delay');
+const delay = require('../../../lib/knex-builder/internal/delay');
 
 module.exports = function (knex) {
   // Certain dialects do not have proper insert with returning, so if this is true
